requirements.txt
openai
PyPDF2
python-dotenv
tiktoken
requests